refactor: fix imports in combat behaviors
