+++ src/ares/behaviors/combat/individual/shoot_and_move_to_target.py
@@ -6,10 +6,11 @@ from sc2.position import Point2
 from sc2.unit import Unit
 from sc2.units import Units
 
-from ares.behaviors.combat.individual import PathUnitToTarget, ShootTargetInRange
+from ares.behaviors.combat.individual import PathUnitToTarget
 from ares.behaviors.combat.individual.combat_individual_behavior import (
     CombatIndividualBehavior,
 )
+from ares.behaviors.combat.individual.shoot_target_in_range import ShootTargetInRange
 from ares.managers.manager_mediator import ManagerMediator
 
 if TYPE_CHECKING:
